Added code to prompt the user before resetting the database to help avoid mistakes.

+++ lightcurve_pipeline/database/reset_database.py
@@ -5,8 +5,16 @@ Reset all tables in the database.
 """
 
 from lightcurve_pipeline.database.database_interface import base
+from lightcurve_pipeline.settings.settings import SETTINGS
 
 if __name__ == '__main__':
 
-    base.metadata.drop_all()
-    base.metadata.create_all()
+    prompt = 'About to reset database instance {}. '.format(SETTINGS['db_connection_string'])
+    prompt += 'Do you wish to proceed? (y/n)'
+
+    response = raw_input(prompt)
+
+    if response.lower() == 'y':
+        print 'Resetting database'
+        base.metadata.drop_all()
+        base.metadata.create_all()
